refactorice code from Grid screen
