+++ src/components/atoms/GeneralText/styled.js
@@ -4,4 +4,5 @@ export const AppText = styled.Text`
   font-size: ${props => props.size};
   color: ${props => props.color};
   font-weight: ${props => (props.weight ? 'bold' : 'normal')};
+  align-self: center;
 `;
